+++ GUI-dev/src/app/@theme/components/auth/jwt.interceptor.ts
@@ -1,0 +1,28 @@
+import 'rxjs/add/operator/do';
+import {
+  HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest,
+  HttpResponse,
+} from '@angular/common/http';
+import {NbAuthService} from '@nebular/auth';
+import {Observable} from 'rxjs/Observable';
+
+export class JwtInterceptor implements HttpInterceptor {
+
+  constructor(public auth: NbAuthService) {}
+
+  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+
+    return next.handle(request).do((event: HttpEvent<any>) => {
+      if (event instanceof HttpResponse) {
+        // do stuff with response if you want
+      }
+    }, (err: any) => {
+      if (err instanceof HttpErrorResponse) {
+        if (err.status === 401) {
+          // redirect to the login route
+          // or show a modal
+        }
+      }
+    });
+  }
+}