+++ GUI-dev/src/app/@theme/providers/backend-server/serverconfig.ts
@@ -4,11 +4,11 @@ import { Injectable } from '@angular/core';
 @Injectable()
 export class ServerConfigProvider {
 
-  private _host = 'http://localhost:10000';
-  private _checkIfMailTaken = `${this._host}/user/register/checkIfMailTaken/`;
-
+  private _host = 'http://localhost:10000/';
+  private _checkIfMailTaken = `${this._host}user/register/checkIfMailTaken/`;
+  private _getUser = `${this._host}/api/me/`;
 
   public get checkIfMailTaken(): string {return this._checkIfMailTaken};
-
+  public get getUser(): string {return this._getUser};
 
 }