+++ GUI-dev/src/app/@theme/theme.module.ts
@@ -44,6 +44,10 @@ import {
 } from './components/auth';
 import { EbEmailPassAuthProvider } from './providers/auth/email-pass-auth.provider';
 import { EbTacComponent } from './components/tac/tac.component';
+import {GatewayProvider} from './providers/backend-server/gateway';
+import {ServerConfigProvider} from './providers/backend-server/serverconfig';
+import {HTTP_INTERCEPTORS} from '@angular/common/http';
+import {TokenInterceptor} from './components/auth/token.interceptor';
 
 const BASE_MODULES = [CommonModule, FormsModule, ReactiveFormsModule];
 
@@ -93,6 +97,15 @@ const PIPES = [
 
 const NB_THEME_PROVIDERS = [
   EbEmailPassAuthProvider,
+  GatewayProvider,
+  ServerConfigProvider,
+  [
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: TokenInterceptor,
+      multi: true,
+    },
+  ],
   ...NbThemeModule.forRoot(
     {
       name: 'default',