+++ GUI-dev/src/app/pages/pages.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
-
 import { PagesComponent } from './pages.component';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { PagesRoutingModule } from './pages-routing.module';
 import { ThemeModule } from '../@theme/theme.module';
 import {HomeModule} from './home/home.module';
+import {UserDetailsModule} from './user-details/user-details.module';
 
 const PAGES_COMPONENTS = [
   PagesComponent,
@@ -16,6 +16,8 @@ const PAGES_COMPONENTS = [
     ThemeModule,
     DashboardModule,
     HomeModule,
+    UserDetailsModule,
+
   ],
   declarations: [
     ...PAGES_COMPONENTS,