+++ GUI-dev/src/app/pages/user-details/user-details.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { ThemeModule } from '../../@theme/theme.module';
+import {UserDetailsComponent} from './user-details.component';
+
+@NgModule({
+  imports: [
+    ThemeModule,
+  ],
+  declarations: [
+    UserDetailsComponent,
+  ],
+})
+export class UserDetailsModule { }