+++ GUI-dev/src/models/models.ts
@@ -12,6 +12,28 @@ export class User {
   username: string;
   email: string;
   password: string;
+  createdAt: Date;
+  organization?: Organization;
+  roles: [Role];
+}
+
+export class Organization {
+  organizationName: string;
+  description: string;
+  oid: number;
+}
+
+export class Role {
+  roleId: number;
+  name: string;
+  systemId: string;
+  rules: [Rule];
+}
+
+export class Rule {
+  ruleId: number;
+  name: string;
+  systemId: string;
 }